added part of auth logic

+++ Src/Auth/WebForum.Auth.Application/Commands/DeleteUserCommand.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using WebForum.Auth.Application.Interfaces;
+using WebForum.Auth.Domain.Exceptions;
 
 namespace WebForum.Auth.Application.Commands;
 
@@ -13,6 +14,11 @@ internal sealed class DeleteUserCommandHandler(
 {
     public async Task Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
+        if (await userRepository.FindByUserId(request.UserId, cancellationToken) is null)
+        {
+            throw new UserNotFoundException(request.UserId);
+        }
+        
         await userRepository.Delete(request.UserId, cancellationToken);
     }
 }
